+++ core/shared/src/main/scala/fs2/async/async.scala
@@ -21,17 +21,18 @@ package object async {
   def semaphore[F[_]: Concurrent](initialCount: Long): F[mutable.Semaphore[F]] =
     mutable.Semaphore(initialCount)
 
-  /** Creates an unbounded asynchronous queue. See [[mutable.Queue]] for more documentation. */
-  def unboundedQueue[F[_]: Concurrent, A](implicit ec: ExecutionContext): F[mutable.Queue[F, A]] =
-    mutable.Queue.unbounded[F, A]
+  /** Creates an unbounded asynchronous queue. See [[mutable.SignallingQueue]] for more documentation. */
+  def unboundedQueue[F[_]: Concurrent, A](
+      implicit ec: ExecutionContext): F[mutable.SignallingQueue[F, A]] =
+    mutable.SignallingQueue.unbounded[F, A]
 
   /**
     * Creates a bounded asynchronous queue. Calls to `enqueue1` will wait until the
-    * queue's size is less than `maxSize`. See [[mutable.Queue]] for more documentation.
+    * queue's size is less than `maxSize`. See [[mutable.SignallingQueue]] for more documentation.
     */
   def boundedQueue[F[_]: Concurrent, A](maxSize: Int)(
-      implicit ec: ExecutionContext): F[mutable.Queue[F, A]] =
-    mutable.Queue.bounded[F, A](maxSize)
+      implicit ec: ExecutionContext): F[mutable.SignallingQueue[F, A]] =
+    mutable.SignallingQueue.bounded[F, A](maxSize)
 
   /**
     * Creates a synchronous queue, which always has size 0. Any calls to `enqueue1`
@@ -39,8 +40,8 @@ package object async {
     * block until there is an offsetting call to `enqueue1`.
     */
   def synchronousQueue[F[_], A](implicit F: Concurrent[F],
-                                ec: ExecutionContext): F[mutable.Queue[F, A]] =
-    mutable.Queue.synchronous[F, A]
+                                ec: ExecutionContext): F[mutable.SignallingQueue[F, A]] =
+    mutable.SignallingQueue.synchronous[F, A]
 
   /**
     * Creates a queue that functions as a circular buffer. Up to `size` elements of
@@ -48,9 +49,10 @@ package object async {
     * the oldest elements. Thus an enqueue process will never wait.
     * @param maxSize The size of the circular buffer (must be > 0)
     */
-  def circularBuffer[F[_], A](maxSize: Int)(implicit F: Concurrent[F],
-                                            ec: ExecutionContext): F[mutable.Queue[F, A]] =
-    mutable.Queue.circularBuffer[F, A](maxSize)
+  def circularBuffer[F[_], A](maxSize: Int)(
+      implicit F: Concurrent[F],
+      ec: ExecutionContext): F[mutable.SignallingQueue[F, A]] =
+    mutable.SignallingQueue.circularBuffer[F, A](maxSize)
 
   /**
     * Converts a discrete stream to a signal. Returns a single-element stream.
